backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
httpx==0.27.2
openai-whisper==20240930
groq==0.11.0
reportlab==4.2.5
minio==7.2.9
python-multipart==0.0.12
pydantic==2.9.2